backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.3.1
    # via kombu
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   google-genai
    #   httpx
asgiref==3.9.1
    # via
    #   django
    #   django-allauth
    #   django-cors-headers
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
billiard==4.2.1
    # via celery
black==25.1.0
    # via -r requirements.in
blessed==1.21.0
    # via jurigged
cachetools==5.5.2
    # via google-auth
celery==5.5.3
    # via -r requirements.in
certifi==2025.7.14
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
codefind==0.1.7
    # via jurigged
cryptography==45.0.5
    # via -r requirements.in
django==4.2.23
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-extensions
    #   django-health-check
    #   django-redis
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-allauth==65.10.0
    # via -r requirements.in
django-cors-headers==4.7.0
    # via -r requirements.in
django-debug-toolbar==6.0.0
    # via -r requirements.in
django-environ==0.12.0
    # via -r requirements.in
django-extensions==4.1
    # via -r requirements.in
django-health-check==3.20.0
    # via -r requirements.in
django-injector==0.3.1
    # via -r requirements.in
django-redis==6.0.0
    # via -r requirements.in
djangorestframework==3.16.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.5.1
    # via -r requirements.in
dnspython==2.7.0
    # via pymongo
docstring-parser==0.17.0
    # via google-cloud-aiplatform
drf-spectacular==0.28.0
    # via -r requirements.in
factory-boy==3.3.3
    # via -r requirements.in
faker==37.5.3
    # via factory-boy
flake8==7.3.0
    # via -r requirements.in
google-api-core[grpc]==2.25.1
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-resource-manager
    #   google-cloud-storage
google-auth==2.40.3
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-resource-manager
    #   google-cloud-storage
    #   google-genai
google-cloud-aiplatform==1.106.0
    # via -r requirements.in
google-cloud-bigquery==3.35.1
    # via google-cloud-aiplatform
google-cloud-core==2.4.3
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-resource-manager==1.14.2
    # via google-cloud-aiplatform
google-cloud-storage==2.19.0
    # via google-cloud-aiplatform
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-genai==1.28.0
    # via google-cloud-aiplatform
google-resumable-media==2.7.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos[grpc]==1.70.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.2
    # via google-cloud-resource-manager
grpcio==1.74.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   qdrant-client
grpcio-status==1.74.0
    # via google-api-core
h11==0.16.0
    # via httpcore
h2==4.2.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx[http2]==0.28.1
    # via
    #   google-genai
    #   qdrant-client
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
inflection==0.5.1
    # via drf-spectacular
iniconfig==2.1.0
    # via pytest
injector==0.22.0
    # via django-injector
isort==6.0.1
    # via -r requirements.in
jsonschema==4.25.0
    # via drf-spectacular
jsonschema-specifications==2025.4.1
    # via jsonschema
jurigged==0.6.1
    # via -r requirements.in
kombu==5.5.4
    # via
    #   -r requirements.in
    #   celery
mccabe==0.7.0
    # via flake8
mongoengine==0.29.1
    # via -r requirements.in
mypy-extensions==1.1.0
    # via black
numpy==2.3.2
    # via
    #   qdrant-client
    #   shapely
ovld==0.5.9
    # via jurigged
packaging==25.0
    # via
    #   black
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   kombu
    #   pytest
pathspec==0.12.1
    # via black
pika==1.3.2
    # via -r requirements.in
pillow==11.3.0
    # via -r requirements.in
platformdirs==4.3.8
    # via black
pluggy==1.6.0
    # via pytest
portalocker==3.2.0
    # via qdrant-client
prompt-toolkit==3.0.51
    # via click-repl
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-resource-manager
protobuf==6.31.1
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-resource-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
    #   qdrant-client
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycodestyle==2.14.0
    # via flake8
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via
    #   google-cloud-aiplatform
    #   google-genai
    #   qdrant-client
pydantic-core==2.33.2
    # via pydantic
pyflakes==3.4.0
    # via flake8
pygments==2.19.2
    # via pytest
pyjwt==2.10.1
    # via djangorestframework-simplejwt
pymongo==4.13.2
    # via
    #   -r requirements.in
    #   mongoengine
pytest==8.4.1
    # via
    #   -r requirements.in
    #   pytest-django
pytest-django==4.11.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   celery
    #   google-cloud-bigquery
python-decouple==3.8
    # via -r requirements.in
pyyaml==6.0.2
    # via drf-spectacular
qdrant-client==1.15.1
    # via -r requirements.in
redis==6.2.0
    # via
    #   -r requirements.in
    #   django-redis
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.4
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   google-genai
rpds-py==0.26.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
shapely==2.1.1
    # via google-cloud-aiplatform
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlparse==0.5.3
    # via
    #   django
    #   django-debug-toolbar
tenacity==8.5.0
    # via google-genai
typing-extensions==4.14.1
    # via
    #   google-cloud-aiplatform
    #   google-genai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via
    #   faker
    #   kombu
uritemplate==4.2.0
    # via drf-spectacular
urllib3==2.5.0
    # via
    #   qdrant-client
    #   requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
watchdog==6.0.0
    # via jurigged
wcwidth==0.2.13
    # via
    #   blessed
    #   prompt-toolkit
websockets==15.0.1
    # via google-genai
